// src/app/model/EstimateMarkupsModel.ts
export class EstimateMarkupsModel {
    ConstructionTotal: number;
    MarkupId: number;
    EstimateId: number;
    ParentMarkupId: number;
   // LMarkupTypeId: number;
    MarkupName: string;
    MarkupType: string;
    LAdjustmentTypeId: number;
    AdjustmentType: string;
    AdjustmentValue: number;
    AdjustmentTotal: number;
    DisplayOrder: number;
    MarkupTotal: number;

}